third_party/blink/web_tests/external/wpt/fetch/api/abort/keepalive-expected.txt

This is a testharness.js-based test.
FAIL aborting a keepalive fetch should work assert_equals: connection should be open expected "open" but got "closed"
PASS aborting a detached keepalive fetch should not do anything
Harness: the test ran to completion.
